Rename beacon header and cpp
